Adjust multiple summary transforms in shared class.
