+++ config-model/src/main/java/com/yahoo/schema/processing/Processing.java
@@ -53,8 +53,8 @@ public class Processing {
                 ImplicitSummaries::new,
                 ImplicitSummaryFields::new,
                 AdjustPositionSummaryFields::new,
-                SummaryTransformForDocumentId::new,
                 SummaryConsistency::new,
+                AdjustSummaryTransforms::new,
                 SummaryNamesFieldCollisions::new,
                 SummaryFieldsMustHaveValidSource::new,
                 MatchedElementsOnlyResolver::new,
